reverse_job_board/requirements.txt
dj-database-url==0.4.1
Django==1.10.4
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
Pillow==4.1.0
requests==2.14.2
beautifulsoup4==4.6.0
django-allauth==0.32.0
